Add per-component debounce and setTimeout
This allows us to do the "when you stop typing for 500 ms, flush edits
to PouchDB" logic, but also flush any pending edits immediately when the
component unmounts.

This fixes a bug I had described as:

> Stop typing in a note, then quickly return to folder list. The
  summary of the note is stale, and never updates. I need some way to
  keep track of pending actions and flush before leaving edit mode

Maybe this is overengineered, there's a better solution, etc., not sure.
The way the debounced function is created (in componentWillMount) is a
bit ugly and seems fragile, as a FIXME comment notes. For now it works.